moved MapBackgroundPlaceholder to molecules
