+++ src/components/molecules/MapBackgroundPlaceholder/MapBackgroundPlaceholder.scss
