+++ src/components/molecules/MapBackgroundPlaceholder/MapBackgroundPlaceholder.tsx
