+++ src/components/molecules/MapBackgroundPlaceholder/index.ts
