+++ src/pages/Admin/MapPreview/MapPreview.tsx
@@ -21,7 +21,7 @@ import {
   SubVenueIconMap,
 } from "pages/Account/Venue/VenueMapEdition/Container";
 
-import { MapBackgroundPlaceholder } from "components/organisms/AdminVenueView/components/MapBackgroundPlaceholder";
+import { MapBackgroundPlaceholder } from "components/molecules/MapBackgroundPlaceholder";
 
 import { ButtonNG } from "components/atoms/ButtonNG/ButtonNG";
 import Legend from "components/atoms/Legend";
